Launch Intents from QS security footer as current user
Intents triggered from QS should be resolved as the current user,
not as the user that QS is running as (which will always be user 0).
ActivityStarter knows how to do this and can handle this for us
automatically.

Bug: 36118203
Test: Manual

Change-Id: I48d880f0a17d7a9f218ab07ef807ad49ebeda624
(cherry picked from commit 6ae2fce2c0abec70e9d9c44b45a917824f7099e9)