Change wifiScanActualTime csv from usec to msec
The wifi scan actual time is desired (in the batterystats.proto) to be
in msec instead of usec. It makes more sense to therefore output the
value in msec rather than convert it when it gets read into the proto.

Bug: 35677312
Test: runtest -x
frameworks/base/core/tests/coretests/src/com/android/internal/os/BatteryStatsTests.java

Change-Id: I22fce69b6faecd00f79837369344e04b426a1c33