Add logging to WebView zygote package preload.
This is a potentially slow event and having a timestamp for when it
started/completed may help to track down a currently-open bug (as well
as enable us to see how long preloading took). This only happens once
per boot and once per WebView update so won't cause a significant amount
of logging.

Bug: 37347817
Test: check for messages in logcat on first use of WebView
Change-Id: Ib6c50eea71324f81cda5b5b1ac91f9dc806e0ab4