Reset touch location after touch slop is exceeded in recents
There was a jump when a user touch moves over the slop value. To fix
this the location is updated to new y position so that the move
animation of the card after has no visual jump.

Change-Id: Ia9b1d04cc0b2983c2e61114ebbf3bdeb154404e9
Fixes: 35720341
Test: manual - enter recents and slowly drag a card vertically and
notice no jump with the fix, moving after the slop has a consistent
motion