Address ConnectivityManager API comments.
This patch is a cherry pick of the two following commits:
  - 024c649372804cde9fa68d8811462465ddc67aca which addresses several
    issues in the public api of ConnectivityManager.
  - 94cc2feb976ebf6a096c39598d2000b09ec6a354 which fixes the documentation
    of several methods in ConnectivityManager public api.

Because the first commit change the public api that is referenced in
the documentation fixed by the second commit, it is not possible to one
without the other. In both cases trying to cherry pick only one of them
results in a build error.

The first commit was submitted successfully on an internal branch before
the checks done in the built got stricter.

Bug: 36370941
Test: marlin builds and boots
Change-Id: I86dcf056e6b165e527c3ee88dbabc2764ac09a08
Merged-In: I693ee5270bf186c88c7c5056293519f7237504ff

(cherry picked from commit 024c649372804cde9fa68d8811462465ddc67aca)