Increase the timeout and add stopping managed profile test
The dashbaord suggests that multiple tests are always capped at the same
value due to the timeout. To collect the actual metric, set a larger
timeout value, 10s for now.

Also, added a test to stop managed profile user. However, the test is
still finishing surprisingly fast because of the empty broadcast queue.


Bug: 62096041

Test: adb shell am instrument -e class android.multiuser.UserLifecycleTest -w com.android.perftests.multiuser/android.support.test.runner.AndroidJUnitRunner

Change-Id: Ia52db17525d9eb960c9a8c860cdb52c0f487e5f1